Changed who the voting system @

+++ cogs/Vote.py
@@ -3,6 +3,7 @@ import discord
 from discord.ext import commands
 import datetime
 
+
 class Vote(commands.Cog):
     def __init__(self, client):
         self.client = client
@@ -12,15 +13,15 @@ class Vote(commands.Cog):
         global message, yes, no
         mod = discord.utils.get(ctx.guild.roles, name="Mods")  # Gets the role "Mod" from the server
         admin = discord.utils.get(ctx.guild.roles, name=":)")  # Gets the role ":)" from the server
-        if mod in ctx.author.roles or admin in ctx.author.roles:
-            converted = str(datetime.timedelta(seconds= int(duration)))
+        if mod in ctx.author.roles or admin in ctx.author.roles and time == True:
+            converted = str(datetime.timedelta(seconds=int(duration)))
             await ctx.message.delete()
             print("Done")
             output = False
             embed = discord.Embed(title=f"Voting. (Duration {converted})")
-            #embed.set_author(name=ctx.message.display_name, url=discord.Embed.Empty, icon_url=ctx.author.avatar_url)
+            # embed.set_author(name=ctx.message.display_name, url=discord.Embed.Empty, icon_url=ctx.author.avatar_url)
             embed.add_field(name="Please vote on", value=vote)
-            message = await ctx.send("@everyone", embed=embed)
+            message = await ctx.send("<@&865011222898016287>", embed=embed)
             await message.add_reaction('✅')
             await message.add_reaction("❌")
             await asyncio.sleep(int(duration))
